+++ Asset Management Platform/Trade.cs
@@ -14,11 +14,11 @@ namespace Asset_Management_Platform
         public string Ticker;
         public int Shares;
         public string Terms;
-        public double Limit;
+        public decimal Limit;
         public string OrderDuration;
 
 
-        public Trade(string buyOrSell, Security securityToAdd, string ticker, int shares, string terms, double limit, string orderDuration)
+        public Trade(string buyOrSell, Security securityToAdd, string ticker, int shares, string terms, decimal limit, string orderDuration)
         {
             BuyOrSell = buyOrSell;
             Security = securityToAdd;
